Simplify library usage: offload the view to another thread
Previously the view would run on the main thread, blocking any possible other actions. In addition, the implementation part which dealt with closing the window was complicated to use.

This changeset pushes the view to its own thread and lets the user actively send an event to close the window. Alternatively, a user may decide to wait until the window closes (or the object stops to live, whichever comes first).

Thanks for the excellent feedback @github.com/kgrzywocz