cargo-deny: Ignore RUSTSEC-2021-0119 for now
Depends on the next release of winit (> 0.25.0), and winit_input_helper (> 0.10.0). If necessary we could replace or fork winit_input_helper if winit is released first and winit_input_helper would not be updated.

The scope, user base and risks of using this small library do not weigh up against ignoring the advisory until a fix becomes available. Fixing it would be preferable of course.